ADD gap finding capability
Flyappy is now able to evaluate the measurements that have been taken up to this point and locate the y-coordinate with the highest chance of being the gap. This is done with the new findGap() function which returns a new gap struct containing a y coordinate and a quality parameter which can later be used for better path planning. Currently tested manually and not working, next step is writing tests and debugging.