CHANGED x and y accel in several small ways
Added some features to both accelerations, added the option to accelerate and decelerate in the pipe if there is time for it (this wasn't very effective) added a confidence rating that limits x-velocities based on gap quality ratings and finally added early accelerator offset which allows y-accel while still in the pipe. The last one makes a big change but it increases the possibility of Flyappy going the wrong way, and also of crashes. Best score: 84